-Add logic to record task of user

+++ PostCelebMoment/src/main/java/com/manabie/postcelebmoment/PostCelebMomentApplication.java
@@ -3,7 +3,10 @@ package com.manabie.postcelebmoment;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
+import java.util.LinkedHashMap;
 
 import javax.annotation.PostConstruct;
 
@@ -25,7 +28,8 @@ import com.amazonaws.services.s3.model.PutObjectRequest;
 public class PostCelebMomentApplication {
 	
 	private AmazonS3 s3client;
-
+	private LinkedHashMap<String, UserTracking> userTrackList = new LinkedHashMap<String, UserTracking>();
+	
 	@Value("${amazonProperties.endpointUrl}")
 	private String endpointUrl;
 	@Value("${amazonProperties.bucketName}")
@@ -41,8 +45,8 @@ public class PostCelebMomentApplication {
 	private void initializeAmazon() {
 		BasicAWSCredentials creds = new BasicAWSCredentials(this.accessKey, this.secretKey); 
 		this.s3client = AmazonS3ClientBuilder.standard().withCredentials(new AWSStaticCredentialsProvider(creds)).withRegion(this.region).build();
+		this.userTrackList = CSVHelper.csvToUserTracking();
 	}
-	
 	
 	private File convertMultiPartToFile(MultipartFile file) throws IOException {
 	    File convFile = new File(file.getOriginalFilename());
@@ -67,16 +71,50 @@ public class PostCelebMomentApplication {
 	        return false;
 	      }
 	}
+
+	private boolean isExceedLimit(String userId) {
+		DateTimeFormatter dtf = DateTimeFormatter.ofPattern("dd-MMM-yyyy HH:mm");  
+		LocalDateTime today = LocalDateTime.now();
+		//insert new user incase user not existed
+		if(!this.userTrackList.containsKey(userId)) {
+			UserTracking newUser = new UserTracking(userId, today, 5, 1);
+			this.userTrackList.put(userId, newUser);
+			return false;
+		}else {
+			//compare lastPostDate and counter
+			LocalDateTime lastPostDate = this.userTrackList.get(userId).getLastPostDate();
+			if(lastPostDate.getDayOfYear() < today.getDayOfYear() || lastPostDate.getYear() < today.getYear()) {
+				//reset counter for first post on day
+				this.userTrackList.get(userId).setCounter(1);
+				this.userTrackList.get(userId).setLastPostDate(today);
+				return false;
+			}else {
+				int counter = this.userTrackList.get(userId).getCounter();
+				if(counter + 1 <= this.userTrackList.get(userId).getLimitation()) {
+					//counter + 1
+					this.userTrackList.get(userId).setCounter(counter + 1);
+					return false;
+				}else {
+					return true;
+				}
+			}
+		}
+	}
 	
-	public String uploadFile(MultipartFile multipartFile) {
+	public String uploadFile(MultipartFile multipartFile, String userId) {
 
 	    String fileUrl = "";
 	    try {
+	    	//TODO validate userId
+	    	if(isExceedLimit(userId)) {
+	    		return "You only can post " + this.userTrackList.get(userId).getLimitation() + " photos per day";
+	    	}
 	        File file = convertMultiPartToFile(multipartFile);
 	        String fileName = generateFileName(multipartFile);
 	        fileUrl = endpointUrl + "/" + bucketName + "/" + fileName;
 	        uploadFileTos3bucket(fileName, file);
 	        file.delete();
+	        CSVHelper.userTrackingsToCSV(this.userTrackList);
 	    } catch (Exception e) {
 	       e.printStackTrace();
 	       return "";
@@ -140,7 +178,7 @@ public class PostCelebMomentApplication {
 	public void setS3client(AmazonS3 s3client) {
 		this.s3client = s3client;
 	}
-
+	
 	public static void main(String[] args) {
 		SpringApplication.run(PostCelebMomentApplication.class, args);
 	}
